Put tests for constrained priors into their own file
